Move test files for easier merging
